requirements.txt
SQLAlchemy~=2.0.39
fastapi~=0.115.11
pydantic~=2.10.6
PyJWT~=2.10.1
passlib~=1.7.4